+++ src/components/myPage/main/MemberSummary.js
@@ -6,6 +6,7 @@ import {
 } from '../../../context/profile.context';
 
 import { Link } from 'react-router-dom';
+import { toCurrencyString } from '../../../utils/unit';
 
 const memberGradeClassName = {
   membership: 'family',
@@ -14,21 +15,7 @@ const memberGradeClassName = {
 
 };
 
-const MemberSummary = ({ tabChange }) => {
-  const { my, profile } = useProfileState();
-  const profileDispatch = useProileDispatch();
-
-  useEffect(() => {
-    if (!my && profile?.memberId) {
-      fetchMy(profile.memberId);
-    }
-  }, [my, profile]);
-
-  function fetchMy (customerid) {
-    fetchMyProfile(profileDispatch, { type: '30', customerid }).
-      catch(console.error);
-  }
-
+const MemberSummary = ({ tabChange, profile, availablemileage }) => {
   const gradeClassName = useMemo(() => {
     const grade = profile?.memberGradeName?.toLowerCase();
     const key = Object.keys(memberGradeClassName).find(k => k === grade);
@@ -55,7 +42,7 @@ const MemberSummary = ({ tabChange }) => {
               <span className="val_txt">
                             <span
                               className={gradeClassName}>{profile?.memberGradeName ||
-                            ''}</span>{/* class: 별 등급 색상 지정 vvip / vip / family */}
+                            ''}</span>
                         </span>
             </Link>
           </li>
@@ -65,7 +52,8 @@ const MemberSummary = ({ tabChange }) => {
                       <span className="ico_txt"><span
                         className="txt_arrow">마일리지</span></span>
               <span className="val_txt"><span
-                className="val">N</span>M</span>
+                className="val">{toCurrencyString(
+                availablemileage)}</span>M</span>
             </a>
           </li>
           <li className="user_item coupon">
